woocommerce_pdf/requirements.txt
WeasyPrint>=63.0
Jinja2==3.1.4
jdatetime==5.0.0
python-dotenv==1.0.1
Flask==3.0.2
WooCommerce==3.0.0
requests==2.31.0
